services/auth/requirements.txt
astroid==3.3.5
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dill==0.3.9
Flask==3.0.3
Flask-Cors==5.0.0
Flask-MySQLdb==2.0.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==3.0.2
mccabe==0.7.0
mysqlclient
parso==0.8.4
pip==24.2
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.9.0
pylint==3.3.1
python-dotenv==1.0.1
tomlkit==0.13.2
Werkzeug==3.0.4
pytest
